ADD error widget code refactor
